scripts/client-sync/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sync-ai-configs"
version = "0.1.0"
description = "Sync AI configs (agents, skills, MCP servers, client config) to Codex, Cursor, Gemini."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
dependencies = [
    "onepassword-sdk>=0.4.0",
    "pyyaml>=6.0",
    "pydantic>=2,<3",
    "tomli>=2.0",
    "tomli-w>=1.0",
    "rich>=13.0",
    "questionary>=2.0",
    "pytimeparse>=1.1",
]

[project.optional-dependencies]
dev = ["pytest>=7.0"]

[project.scripts]
sync-ai-configs = "sync_ai_configs.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["sync_ai_configs*"]
exclude = ["tests*", "tests.*", "*.tests*", "*.tests.*"]
